Get station metadata

Returns metadata about measurement stations in the SEPA network.

Usage

sepa_station_list(
  station_search_term,
  bounding_box,
  group_id,
  return_fields,
  ...
)

Arguments

station_search_term

Character. A string pattern to search station names. This argument supports wildcard matching using the asterisk (*) symbol. The search is case insensitive.

bounding_box

Numeric. A bounding box within which to search for stations. This argument should be a vector with exactly four values formatted as follows: (min_x, min_y, max_x, max_y).

group_id

Character or numeric. A station group id. See sepa_group_list().

return_fields

Character. A vector of return fields. In most cases the default return fields will suffice. See details.

...

Additional arguments. None implemented.

Value

A tibble.

Return fields

The argument return_fields allows you to specify the return fields. The default return fields are as follows:

station_name

station_no

station_id

station_latitude

station_longitude

There are too many options to document here. Instead, take a look at the online documentation for getStationList to see the possible values that can be supplied to the API parameter returnfields.

Examples

if (FALSE) { # \dontrun{
# Filter by name
stns <- sepa_station_list(station_search_term="Auch*")

# Filter by region
stns <- sepa_station_list(bounding_box = c(-2, 48, 2, 60))

# Filter by group
q_grp <- grps |>
  dplyr::filter(group_name %in% "StationsWithFlow") |>
  dplyr::pull(group_id)
stns <- sepa_station_list(group_id = q_grp)

# Change the return fields
fields <- c("station_name", "station_no", "station_id", "object_type")
stns <- sepq_station_list(group_id = q_grp, return_fields = fields)
} # }