Get timeseries data for one or more stations

Get timeseries data for one or more stations

Usage

sepa_timeseries_values(
  ts_id,
  start_date,
  end_date,
  return_fields,
  metadata = FALSE,
  md_return_fields,
  ca_sta = FALSE,
  ca_sta_return_fields,
  ...
)

Arguments

ts_id

Character or numeric. Either a single station ID or a vector of station IDs. Time series IDs can be found using the sepa_timeseries_list() function.

start_date

Date or character formatted "YYYY-MM-DD". Defaults to yesterday.

end_date

Date or character formatted "YYYY-MM-DD". Defaults to today.

return_fields

Character. A vector of return fields. In most cases the default return fields will suffice. See details.

metadata

Logical. Whether or not to return metadata alongside the timeseries values.

md_return_fields

Character. A vector of metadata return fields. Only used if metadata is TRUE. See details.

ca_sta

Logical. Whether or not to return custom attribute fields. Default is FALSE.

ca_sta_return_fields

Character.

...

Additional arguments. None implemented.

Value

A tibble.

Return fields

The argument return_fields allows you to specify the return fields. The default return fields are as follows:

Timestamp

Value

It is also possible to return some metadata fields. You can do this by setting metadata=TRUE and supplying md_return_fields. If the latter is not given then the following default fields will be requested:

ts_unitname

ts_unitsymbol

ts_name

ts_id

stationparameter_name

station_name

station_id

Lastly, there are some additional optional return fields that can be returned by setting custom_attribute_fields=TRUE and supplying ca_sta_returnfields. If the latter is not set then the following fields will be requested:

CATCHMENT_SIZE

GAUGE_DATUM

In all cases, note that some fields are not available for some timeseries.

There are too many options to document here. Instead, take a look at the online documentation for getTimeseriesValues to see the possible values that can be supplied to the API parameters returnfields, md_returnfields. Unfortunately the possible arguments to ca_sta_returnfields are currently undocumented.